+++ medaka/__init__.py
@@ -5,7 +5,7 @@ import os
 import subprocess
 import sys
 
-__version__ = "1.6.1"
+__version__ = "1.7.0"
 
 try:
     import parasail